fix: extract checkCompleteness to validation.ts, fix biome import ordering (#206)
- Extract checkCompleteness from fold.ts (310 lines) into validation.ts
  to stay under 300-line limit after biome formatted overloads multi-line
- Fix biome import ordering (type imports before value imports) across
  all plugin test files
