+++ packages/plugin-anthropic/tests/0.74.0/interpreter.test.ts
@@ -1,5 +1,5 @@
-import { coreInterpreter, foldAST, injectInput, mvfm, num, str } from "@mvfm/core";
 import type { Program } from "@mvfm/core";
+import { coreInterpreter, foldAST, injectInput, mvfm, num, str } from "@mvfm/core";
 import { describe, expect, it, vi } from "vitest";
 import { anthropicInterpreter } from "../../src";
 import { anthropic } from "../../src/0.74.0";
