+++ packages/plugin-console/tests/22.0.0/interpreter.test.ts
@@ -1,5 +1,5 @@
-import { coreInterpreter, foldAST, injectInput, mvfm, num, str } from "@mvfm/core";
 import type { Program } from "@mvfm/core";
+import { coreInterpreter, foldAST, injectInput, mvfm, num, str } from "@mvfm/core";
 import { describe, expect, it } from "vitest";
 import { consoleInterpreter } from "../../src";
 import type { ConsoleMethodName } from "../../src/22.0.0";
