+++ packages/plugin-pino/tests/10.3.1/interpreter.test.ts
@@ -1,5 +1,5 @@
-import { coreInterpreter, foldAST, injectInput, mvfm, num, str } from "@mvfm/core";
 import type { Program } from "@mvfm/core";
+import { coreInterpreter, foldAST, injectInput, mvfm, num, str } from "@mvfm/core";
 import { describe, expect, it } from "vitest";
 import { pinoInterpreter } from "../../src";
 import { pino } from "../../src/10.3.1";
