+++ packages/plugin-s3/tests/3.989.0/interpreter.test.ts
@@ -1,5 +1,5 @@
-import { coreInterpreter, foldAST, injectInput, mvfm, num, str } from "@mvfm/core";
 import type { Program } from "@mvfm/core";
+import { coreInterpreter, foldAST, injectInput, mvfm, num, str } from "@mvfm/core";
 import { describe, expect, it } from "vitest";
 import { s3 } from "../../src/3.989.0";
 import { createS3Interpreter, type S3Client } from "../../src/3.989.0/interpreter";
